+++ client-java/src/test/java/io/zeebe/client/workflow/CancelWorkflowInstanceTest.java
@@ -77,9 +77,9 @@ public class CancelWorkflowInstanceTest {
     assertThat(commandRequest.getCommand())
         .containsOnly(
             entry("bpmnProcessId", baseEvent.getBpmnProcessId()),
-            entry("version", baseEvent.getVersion()),
-            entry("workflowKey", (int) baseEvent.getWorkflowKey()),
-            entry("workflowInstanceKey", (int) baseEvent.getWorkflowInstanceKey()),
+            entry("version", (long) baseEvent.getVersion()),
+            entry("workflowKey", baseEvent.getWorkflowKey()),
+            entry("workflowInstanceKey", baseEvent.getWorkflowInstanceKey()),
             entry("activityId", baseEvent.getActivityId()),
             entry("payload", baseEvent.getPayloadField().getMsgPack()));
